Changed the language grammar
to better fit the needs of the semantical analysis later, and to
allow more expressiveness.

Lambdas are a thing now.

Use of control expressions can be done at the deepest level of
operators.

The precedence of operators has been updated.

The associativity of the exponential operator changed from left to
right.